desafio concluido

+++ aula10.18/desafio_da_aula10.18/desafio04.py
@@ -3,6 +3,6 @@ nota2 = input ('informe a nota 2')
 nota3 = input ('informe a nota 3')
 nota4 = input ('informe a nota 4')
 
-media_das_notas = (int(nota1) + int(nota2) + int(nota3) + int(nota4) )/4
+media_das_notas = (float(nota1) + float(nota2) + float(nota3) + float(nota4) )/4
 
 print ("A media das notas eh:", media_das_notas)
